requirements.txt
asttokens==3.0.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
Flask==3.1.2
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
Pygments==2.19.2
PySimpleGUI==5.0.10
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.1.0
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
